Move file into editor package
