+++ packages/block-editor/src/autocompleters/style.scss
@@ -11,11 +11,3 @@
 		color: inherit !important;
 	}
 }
-
-.block-editor-autocompleters__link {
-	white-space: nowrap;
-
-	.block-editor-block-icon {
-		margin-right: $grid-unit-10;
-	}
-}
