+++ packages/editor/src/components/autocompleters/index.js
@@ -1,1 +1,2 @@
+export { default as linkAutocompleter } from './link';
 export { default as userAutocompleter } from './user';
