+++ packages/editor/src/components/autocompleters/style.scss
@@ -1,4 +1,13 @@
 @use "@wordpress/base-styles/colors" as *;
+@use "@wordpress/base-styles/variables" as *;
+
+.editor-autocompleters__link {
+	white-space: nowrap;
+
+	.block-editor-block-icon {
+		margin-right: $grid-unit-10;
+	}
+}
 
 .editor-autocompleters__user {
 	.editor-autocompleters__no-avatar::before {
